Don't inline and unroll multiply/divide in the OS
This is an opportunity for optimization; the included OS is the biggest chunk of code for compilers and translators to work with, and should reflect what "typical" code might look like.

To get good performace in the simulator, plan to add a similar optimization pass as a feature of the runtime.