Line editing
The first call to getchar captures an entire line of input, then pushes the characters one at a time.

This makes the REPL behave in a sensible way, and can handle backspace. It also means that most text entry is always happening when the interpreter is in a type loop handling getchar, so it’s much more responsive.