Use the full available space for heap
And move the invocation of GC to the edge of the interpreter loop so it can’t happen in the middle of any instruction.
That means less overhead in alloc(), but also that we have to leave a fixed margin of space that’s enough for the biggest proc arg list we’re ever going to see.